SMS 2.0
git-svn-id: http://psp.jim.sh/svn/ps2ware/trunk@588 5c23000f-aefb-0310-8550-8024f1b2be0e

+++ include/SMS_PlayerControl.h
@@ -41,7 +41,7 @@ void                 PlayerControl_HandleOSD        ( int, int                  
 unsigned int         PlayerControl_GSPLen           ( struct SMS_List*, unsigned int   );
 unsigned int         PlayerControl_GSPacket         ( int, struct SMS_List*, uint64_t* );
 int                  PlayerControl_ScrollBar        (  void ( * ) ( int )              );
-void                 PlayerControl_UpdateDuration   ( unsigned int, unsigned int       );
+void                 PlayerControl_UpdateDuration   ( unsigned int, int64_t            );
 void                 PlayerControl_UpdateItemNr     ( void                             );
 void                 PlayerControl_AdjustBrightness ( int                              );
 # ifdef __cplusplus
